Removed client-side erdstall dependencies
