+++ client/index.html
@@ -357,9 +357,6 @@ Mozilla presents an HTML5 mini-MMORPG by Little Workshop http://www.littleworksh
         </script>
         <script>var exports = {};</script>
         <script src='/shared/js/erdstallclientinterfacebundle.js'></script>
-        <script>
-            var erdstallInterface = new erdstallClientBundle.default();
-        </script>
         <script data-main="js/home" src="js/lib/require-jquery.js"></script>
 	</body>
 </html>
