+++ ts/erdstallserverinterface.ts
@@ -2,14 +2,16 @@ import { Address } from "@polycrypt/erdstall/ledger";
 import { Asset, Assets, mapNFTs, Tokens } from "@polycrypt/erdstall/ledger/assets";
 import { Session } from "@polycrypt/erdstall";
 import NFT from "./nft";
-import erdstallClientInterface, { getNFTsFromAssets } from "./erdstallclientinterface"
 import { BalanceProof, TxReceipt } from "@polycrypt/erdstall/api/responses";
 import { ethers } from "ethers";
 import config from './config/serverConfig.json';
 import { Burn, Trade, Transfer } from "@polycrypt/erdstall/api/transactions";
 import { Mutex } from "async-mutex";
+import { key } from "./nft";
 
-export default class erdstallServerInterface extends erdstallClientInterface {
+export default class erdstallServerInterface {
+
+	_session: Session | undefined;
 
 	protected nextNftID!: bigint;
 	// Token to mint NFTs on
@@ -147,6 +149,21 @@ export default class erdstallServerInterface extends erdstallClientInterface {
 			}
 		}
 	}
+	
+	// Returns nftKeys belonging to specified or own account
+	async getNFTs(address?: string): Promise< string[] > {
+		if (!this._session) return new Array();
+
+		// Return NFTs belonging to other account if address is specified
+		if (address) {
+			const account = await this._session.getAccount(Address.fromString(address));
+			return getNFTsFromAssets(account.values);
+		}
+		else {
+			const account = await this._session.getOwnAccount();
+			return getNFTsFromAssets(account.values);
+		}
+	}
 
 	// Registers listener function for transfer and burn events
 	registerCallbacks(transferCallback: (sender: string, recipient: string, nfts: string[]) => void, burnCallback: (nfts: string[]) => void) {
@@ -188,4 +205,13 @@ function getAssetsFromNFT(nfts: NFT[]): Assets {
 		})
 	}
 	return new Assets(...assets);
+}
+
+// Extracts string list of nftKeys from assets
+export function getNFTsFromAssets(assets: Assets): string[] {
+	var nfts = new Array();
+	mapNFTs(assets.values, (token, id) => {
+		nfts.push(key(token, id));
+	});
+	return nfts;
 }
